requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
duckdb==1.1.3
APScheduler==3.10.4
requests==2.32.3  
python-dotenv==1.0.0
resend